Moved version logic into common package and updated sidecar and controller to dynamically obtain version
